fix: 회원 도메인 엔티티 값 검증 로직 수정 (#136)
